refactor: unify app.rs into lib.rs and move run() to cli.rs
- Move assembly logic (config_command, daemon_command) to src/lib.rs,
  which now also serves as the [lib] crate root with module declarations
- Add run() routing function to src/cli.rs; each arm delegates to
  the assembly functions in the crate root
- Delete src/app.rs (fully absorbed by lib.rs)
- Update src/main.rs to call into the library crate (merge_ready::cli)
- Add detailed doc comments to config_command and daemon_command
