+++ src/cli.rs
@@ -1,4 +1,6 @@
-use clap::{Parser, Subcommand};
+use std::process::ExitCode;
+
+use clap::{CommandFactory, Parser, Subcommand};
 
 use crate::contexts::daemon::interface::cli::DaemonArgs;
 
@@ -19,4 +21,16 @@ pub enum Command {
     Config,
     /// Manage the background cache daemon
     Daemon(DaemonArgs),
+}
+
+#[allow(clippy::needless_pass_by_value)]
+pub fn run(cli: Cli) -> ExitCode {
+    match cli.command {
+        Some(Command::Config) => crate::config_command(),
+        Some(Command::Daemon(args)) => crate::daemon_command(args),
+        None => {
+            let _ = Cli::command().print_help();
+            ExitCode::SUCCESS
+        }
+    }
 }
